deploy-13.09.2025-1

+++ src/main/java/by/forward/forward_system/core/jpa/repository/projections/OrderChatDataProjection.java
@@ -5,5 +5,5 @@ public interface OrderChatDataProjection {
 
     Long getOrderId();
 
-    Integer getNewMessageCount();
+    Long getNewMessageCount();
 }
